Update 5.2
Added a feature where the view of the priest is determined by the color depending on the upcoming dates or dates that have passed. Also added the full paper regarding this source code and have mended the code into 1 just in case someone wants to read the whole code in 1 file.